sprite.py -> object.py
This is a massive restructuring which moves toward an MVC model by
loosely coupling the components.  The rationale is to ease scene
management and later expansion.  This is primarily a game engine change
(gensokyo).

The first part is revamping the model.  Sprites and Objects are now
separate (view and model).  In order to allow Objects control over their
sprites, a loose coupling is in effect, requiring a global view object
to be available.  Objects extend SpriteWrapper, which is an
EventDispatcher that sends new sprite events to View, which manages the
batch, groups and drawing.