Revamp Scenes
New way is a SceneStack as EventDispatcher.  It will dispatch events to
the MVCs.  Each scene contains an MVC.  Events will be dispatched down
stack until one returns EVENT_HANDLED.  each MVC has a reference to the
stack ("master") so they can dispatch events.

The stack itself has the main window, and registers pushed MVCs that
need it to the window.